Added JSON output option with some refactoring.
